+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 1.3.4
 
+* Azure timestamp parsing is now slightly faster
 * Replace `xmltodict` with `lxml` as it is slightly faster
 * Use lazy import for `pycryptodome` from @hauntsaninja